+++ src/main/java/janggiGame/Board.java
@@ -1,7 +1,6 @@
 package janggiGame;
 
 import janggiGame.arrangement.ArrangementStrategy;
-import janggiGame.board.Dot;
 import janggiGame.piece.Dynasty;
 import janggiGame.piece.Piece;
 import java.util.HashMap;
