+++ src/main/java/view/OutputView.java
@@ -1,6 +1,6 @@
 package view;
 
-import janggiGame.board.Dot;
+import janggiGame.Dot;
 import janggiGame.piece.Dynasty;
 import janggiGame.piece.Piece;
 import java.util.Map;
