+++ src/test/java/janggiGame/board/DotTest.java
@@ -3,6 +3,7 @@ package janggiGame.board;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import janggiGame.Dot;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
