+++ src/test/java/janggiGame/piece/PieceTest.java
@@ -3,7 +3,7 @@ package janggiGame.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import janggiGame.board.Dot;
+import janggiGame.Dot;
 import janggiGame.testhelper.FakePiece;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
