+++ src/test/java/janggiGame/testhelper/FakePiece.java
@@ -1,6 +1,6 @@
 package janggiGame.testhelper;
 
-import janggiGame.board.Dot;
+import janggiGame.Dot;
 import janggiGame.piece.Dynasty;
 import janggiGame.piece.Piece;
 import janggiGame.piece.PieceType;
